# executor,planner: cache the executors' chunks in order to reuse them while executing a plan cached