--- a/executor/benchmark_test.go
+++ b/executor/benchmark_test.go
@@ -127,7 +127,7 @@ func (mds *mockDataSource) Next(ctx context.Context, req *chunk.RecordBatch) err
 }
 
 func buildMockDataSource(opt mockDataSourceParameters) *mockDataSource {
-	baseExec := newBaseExecutor(opt.ctx, opt.schema, "", nil)
+	baseExec := newBaseExecutor(opt.ctx, opt.schema, "", 0, nil)
 	m := &mockDataSource{baseExec, opt, nil, nil, 0}
 	types := m.retTypes()
 	colData := make([][]interface{}, len(types))