--- a/executor/prepared.go
+++ b/executor/prepared.go
@@ -90,7 +90,7 @@ var prepareStmtLabel = stringutil.StringerStr("PrepareStmt")
 
 // NewPrepareExec creates a new PrepareExec.
 func NewPrepareExec(ctx sessionctx.Context, is infoschema.InfoSchema, sqlTxt string) *PrepareExec {
-	base := newBaseExecutor(ctx, nil, prepareStmtLabel)
+	base := newBaseExecutor(ctx, nil, prepareStmtLabel, nil)
 	base.initCap = chunk.ZeroCapacity
 	return &PrepareExec{
 		baseExecutor: base,
@@ -228,6 +228,7 @@ func (e *ExecuteExec) Build(b *executorBuilder) error {
 	if err != nil {
 		return err
 	}
+	b.planKey = e.plan.CacheKey() //HERE: check the caller logic when creating the arg of executorBuilder
 	stmtExec := b.build(e.plan)
 	if b.err != nil {
 		log.Warn("rebuild plan in EXECUTE statement failed", zap.String("labelName of PREPARE statement", e.name))