--- a/executor/prepared.go
+++ b/executor/prepared.go
@@ -90,7 +90,7 @@ var prepareStmtLabel = stringutil.StringerStr("PrepareStmt")
 
 // NewPrepareExec creates a new PrepareExec.
 func NewPrepareExec(ctx sessionctx.Context, is infoschema.InfoSchema, sqlTxt string) *PrepareExec {
-	base := newBaseExecutor(ctx, nil, prepareStmtLabel, nil)
+	base := newBaseExecutor(ctx, nil, prepareStmtLabel, 0, nil)
 	base.initCap = chunk.ZeroCapacity
 	return &PrepareExec{
 		baseExecutor: base,