<!--
Thank you for contributing to TiDB! Please read TiDB's [CONTRIBUTING](https://github.com/pingcap/tidb/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
-->

### What problem does this PR solve? <!--add issue link with summary if exists-->
Fix #9396 

### What is changed and how it works?
* The PR caches the newly created chunks of the executor to the plan cache when executing the plan cached at the first time
* The PR allows the executor to reuse the cached chunks when executing the plan cached from the second time

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

 - Manual test (add detailed scripts or steps below)
```
mysql-host=127.0.0.1
mysql-port=4000
mysql-user=sbtest
mysql-password=pass
mysql-db=sbtest
time=300
threads=64
report-interval=10
db-driver=mysql

sysbench --config-file=config oltp_point_select --tables=32 --table-size=1000000 prepare
sysbench --config-file=config oltp_read_only --tables=32 --table-size=1000000 run
```

Code changes

 - Has exported function/method change
 - Has exported variable/fields change
 - Has interface methods change

Side effects

 - Increased code complexity

Related changes
 - None