@dbjoa 2019-04-11
Here are the benchmark results. In the benchmark, the performance gain of the PR is about 10%, but it might be varied depending on the queries.

[After the PR]
```
% sysbench --config-file=config oltp_read_only --tables=32 --table-size=1000000 run
sysbench 1.0.14-b31eb23 (using bundled LuaJIT 2.1.0-beta3)

Running the test with following options:
Number of threads: 64
Report intermediate results every 10 second(s)
Initializing random number generator from current time


Initializing worker threads...

Threads started!

[ 10s ] thds: 64 tps: 490.83 qps: 7904.43 (r/w/o: 6916.48/0.00/987.95) lat (ms,95%): 193.38 err/s: 0.00 reconn/s: 0.00
[ 20s ] thds: 64 tps: 500.42 qps: 8005.40 (r/w/o: 7004.76/0.00/1000.64) lat (ms,95%): 196.89 err/s: 0.00 reconn/s: 0.00
[ 30s ] thds: 64 tps: 503.80 qps: 8063.00 (r/w/o: 7055.30/0.00/1007.70) lat (ms,95%): 193.38 err/s: 0.00 reconn/s: 0.00
[ 40s ] thds: 64 tps: 512.90 qps: 8205.69 (r/w/o: 7180.79/0.00/1024.90) lat (ms,95%): 189.93 err/s: 0.00 reconn/s: 0.00
[ 50s ] thds: 64 tps: 509.20 qps: 8153.70 (r/w/o: 7134.30/0.00/1019.40) lat (ms,95%): 193.38 err/s: 0.00 reconn/s: 0.00
[ 60s ] thds: 64 tps: 510.00 qps: 8151.80 (r/w/o: 7131.80/0.00/1020.00) lat (ms,95%): 189.93 err/s: 0.00 reconn/s: 0.00
[ 70s ] thds: 64 tps: 513.70 qps: 8227.21 (r/w/o: 7200.31/0.00/1026.90) lat (ms,95%): 189.93 err/s: 0.00 reconn/s: 0.00
[ 80s ] thds: 64 tps: 508.00 qps: 8117.47 (r/w/o: 7100.87/0.00/1016.60) lat (ms,95%): 189.93 err/s: 0.00 reconn/s: 0.00
[ 90s ] thds: 64 tps: 514.80 qps: 8245.13 (r/w/o: 7215.53/0.00/1029.60) lat (ms,95%): 186.54 err/s: 0.00 reconn/s: 0.00
[ 100s ] thds: 64 tps: 512.00 qps: 8178.59 (r/w/o: 7154.99/0.00/1023.60) lat (ms,95%): 193.38 err/s: 0.00 reconn/s: 0.00
[ 110s ] thds: 64 tps: 516.20 qps: 8265.08 (r/w/o: 7232.68/0.00/1032.40) lat (ms,95%): 189.93 err/s: 0.00 reconn/s: 0.00
[ 120s ] thds: 64 tps: 511.10 qps: 8173.13 (r/w/o: 7150.83/0.00/1022.30) lat (ms,95%): 189.93 err/s: 0.00 reconn/s: 0.00
[ 130s ] thds: 64 tps: 513.50 qps: 8218.99 (r/w/o: 7192.09/0.00/1026.90) lat (ms,95%): 189.93 err/s: 0.00 reconn/s: 0.00
[ 140s ] thds: 64 tps: 505.90 qps: 8094.10 (r/w/o: 7082.40/0.00/1011.70) lat (ms,95%): 196.89 err/s: 0.00 reconn/s: 0.00
[ 150s ] thds: 64 tps: 511.27 qps: 8188.40 (r/w/o: 7165.55/0.00/1022.85) lat (ms,95%): 189.93 err/s: 0.00 reconn/s: 0.00
[ 160s ] thds: 64 tps: 519.57 qps: 8303.58 (r/w/o: 7265.43/0.00/1038.15) lat (ms,95%): 186.54 err/s: 0.00 reconn/s: 0.00
[ 170s ] thds: 64 tps: 512.95 qps: 8213.22 (r/w/o: 7186.22/0.00/1027.00) lat (ms,95%): 189.93 err/s: 0.00 reconn/s: 0.00
[ 180s ] thds: 64 tps: 513.10 qps: 8208.26 (r/w/o: 7182.06/0.00/1026.19) lat (ms,95%): 193.38 err/s: 0.00 reconn/s: 0.00
[ 190s ] thds: 64 tps: 512.98 qps: 8211.84 (r/w/o: 7186.17/0.00/1025.67) lat (ms,95%): 193.38 err/s: 0.00 reconn/s: 0.00
[ 200s ] thds: 64 tps: 503.61 qps: 8058.90 (r/w/o: 7052.18/0.00/1006.73) lat (ms,95%): 193.38 err/s: 0.00 reconn/s: 0.00
[ 210s ] thds: 64 tps: 513.79 qps: 8213.56 (r/w/o: 7185.59/0.00/1027.97) lat (ms,95%): 196.89 err/s: 0.00 reconn/s: 0.00
[ 220s ] thds: 64 tps: 514.12 qps: 8231.10 (r/w/o: 7202.36/0.00/1028.74) lat (ms,95%): 193.38 err/s: 0.00 reconn/s: 0.00
[ 230s ] thds: 64 tps: 513.80 qps: 8220.03 (r/w/o: 7192.53/0.00/1027.50) lat (ms,95%): 186.54 err/s: 0.00 reconn/s: 0.00
[ 240s ] thds: 64 tps: 509.99 qps: 8150.87 (r/w/o: 7130.89/0.00/1019.98) lat (ms,95%): 196.89 err/s: 0.00 reconn/s: 0.00
[ 250s ] thds: 64 tps: 513.51 qps: 8229.69 (r/w/o: 7202.77/0.00/1026.92) lat (ms,95%): 189.93 err/s: 0.00 reconn/s: 0.00
[ 260s ] thds: 64 tps: 517.60 qps: 8286.44 (r/w/o: 7251.04/0.00/1035.39) lat (ms,95%): 193.38 err/s: 0.00 reconn/s: 0.00
[ 270s ] thds: 64 tps: 510.87 qps: 8163.36 (r/w/o: 7141.62/0.00/1021.73) lat (ms,95%): 189.93 err/s: 0.00 reconn/s: 0.00
[ 280s ] thds: 64 tps: 523.64 qps: 8380.47 (r/w/o: 7333.70/0.00/1046.77) lat (ms,95%): 186.54 err/s: 0.00 reconn/s: 0.00
[ 290s ] thds: 64 tps: 514.70 qps: 8233.90 (r/w/o: 7204.10/0.00/1029.80) lat (ms,95%): 193.38 err/s: 0.00 reconn/s: 0.00
[ 300s ] thds: 64 tps: 517.70 qps: 8275.70 (r/w/o: 7240.20/0.00/1035.50) lat (ms,95%): 189.93 err/s: 0.00 reconn/s: 0.00
SQL statistics:
    queries performed:
        read:                            2149308
        write:                           0
        other:                           307044
        total:                           2456352
    transactions:                        153522 (511.48 per sec.)
    queries:                             2456352 (8183.68 per sec.)
    ignored errors:                      0      (0.00 per sec.)
    reconnects:                          0      (0.00 per sec.)

General statistics:
    total time:                          300.1501s
    total number of events:              153522

Latency (ms):
         min:                                   44.78
         avg:                                  125.09
         max:                                  311.68
         95th percentile:                      193.38
         sum:                             19204450.80

Threads fairness:
    events (avg/stddev):           2398.7812/7.46
    execution time (avg/stddev):   300.0695/0.04
```
[Before the PR]
```
% sysbench --config-file=config oltp_read_only --tables=32 --table-size=1000000 run
sysbench 1.0.14-b31eb23 (using bundled LuaJIT 2.1.0-beta3)

Running the test with following options:
Number of threads: 64
Report intermediate results every 10 second(s)
Initializing random number generator from current time


Initializing worker threads...

Threads started!

[ 10s ] thds: 64 tps: 436.65 qps: 7036.07 (r/w/o: 6156.97/0.00/879.10) lat (ms,95%): 204.11 err/s: 0.00 reconn/s: 0.00
[ 20s ] thds: 64 tps: 449.72 qps: 7201.07 (r/w/o: 6301.34/0.00/899.73) lat (ms,95%): 200.47 err/s: 0.00 reconn/s: 0.00
[ 30s ] thds: 64 tps: 456.00 qps: 7285.83 (r/w/o: 6374.02/0.00/911.80) lat (ms,95%): 196.89 err/s: 0.00 reconn/s: 0.00
[ 40s ] thds: 64 tps: 465.50 qps: 7459.08 (r/w/o: 6527.58/0.00/931.50) lat (ms,95%): 196.89 err/s: 0.00 reconn/s: 0.00
[ 50s ] thds: 64 tps: 470.40 qps: 7517.59 (r/w/o: 6577.49/0.00/940.10) lat (ms,95%): 196.89 err/s: 0.00 reconn/s: 0.00
[ 60s ] thds: 64 tps: 469.60 qps: 7517.51 (r/w/o: 6578.31/0.00/939.20) lat (ms,95%): 193.38 err/s: 0.00 reconn/s: 0.00
[ 70s ] thds: 64 tps: 459.40 qps: 7356.69 (r/w/o: 6437.29/0.00/919.40) lat (ms,95%): 196.89 err/s: 0.00 reconn/s: 0.00
[ 80s ] thds: 64 tps: 465.20 qps: 7444.51 (r/w/o: 6514.11/0.00/930.40) lat (ms,95%): 196.89 err/s: 0.00 reconn/s: 0.00
[ 90s ] thds: 64 tps: 469.52 qps: 7502.58 (r/w/o: 6563.53/0.00/939.05) lat (ms,95%): 193.38 err/s: 0.00 reconn/s: 0.00
[ 100s ] thds: 64 tps: 461.17 qps: 7384.60 (r/w/o: 6462.15/0.00/922.45) lat (ms,95%): 196.89 err/s: 0.00 reconn/s: 0.00
[ 110s ] thds: 64 tps: 463.89 qps: 7413.52 (r/w/o: 6485.73/0.00/927.79) lat (ms,95%): 196.89 err/s: 0.00 reconn/s: 0.00
[ 120s ] thds: 64 tps: 457.31 qps: 7319.28 (r/w/o: 6404.77/0.00/914.51) lat (ms,95%): 196.89 err/s: 0.00 reconn/s: 0.00
[ 130s ] thds: 64 tps: 468.80 qps: 7502.89 (r/w/o: 6565.49/0.00/937.40) lat (ms,95%): 193.38 err/s: 0.00 reconn/s: 0.00
[ 140s ] thds: 64 tps: 464.50 qps: 7437.21 (r/w/o: 6507.91/0.00/929.30) lat (ms,95%): 196.89 err/s: 0.00 reconn/s: 0.00
[ 150s ] thds: 64 tps: 468.96 qps: 7507.32 (r/w/o: 6570.19/0.00/937.13) lat (ms,95%): 193.38 err/s: 0.00 reconn/s: 0.00
[ 160s ] thds: 64 tps: 464.38 qps: 7420.64 (r/w/o: 6491.28/0.00/929.35) lat (ms,95%): 193.38 err/s: 0.00 reconn/s: 0.00
[ 170s ] thds: 64 tps: 463.65 qps: 7438.09 (r/w/o: 6510.60/0.00/927.50) lat (ms,95%): 196.89 err/s: 0.00 reconn/s: 0.00
[ 180s ] thds: 64 tps: 473.89 qps: 7572.45 (r/w/o: 6624.97/0.00/947.48) lat (ms,95%): 189.93 err/s: 0.00 reconn/s: 0.00
[ 190s ] thds: 64 tps: 464.22 qps: 7419.99 (r/w/o: 6491.25/0.00/928.74) lat (ms,95%): 196.89 err/s: 0.00 reconn/s: 0.00
[ 200s ] thds: 64 tps: 470.19 qps: 7524.22 (r/w/o: 6583.84/0.00/940.38) lat (ms,95%): 189.93 err/s: 0.00 reconn/s: 0.00
[ 210s ] thds: 64 tps: 467.21 qps: 7474.78 (r/w/o: 6540.55/0.00/934.22) lat (ms,95%): 193.38 err/s: 0.00 reconn/s: 0.00
[ 220s ] thds: 64 tps: 466.00 qps: 7446.48 (r/w/o: 6515.49/0.00/931.00) lat (ms,95%): 196.89 err/s: 0.00 reconn/s: 0.00
[ 230s ] thds: 64 tps: 461.50 qps: 7385.52 (r/w/o: 6461.32/0.00/924.20) lat (ms,95%): 196.89 err/s: 0.00 reconn/s: 0.00
[ 240s ] thds: 64 tps: 466.20 qps: 7468.10 (r/w/o: 6535.90/0.00/932.20) lat (ms,95%): 193.38 err/s: 0.00 reconn/s: 0.00
[ 250s ] thds: 64 tps: 468.70 qps: 7489.30 (r/w/o: 6552.60/0.00/936.70) lat (ms,95%): 193.38 err/s: 0.00 reconn/s: 0.00
[ 260s ] thds: 64 tps: 469.60 qps: 7512.38 (r/w/o: 6572.29/0.00/940.10) lat (ms,95%): 193.38 err/s: 0.00 reconn/s: 0.00
[ 270s ] thds: 64 tps: 472.50 qps: 7560.31 (r/w/o: 6615.91/0.00/944.40) lat (ms,95%): 189.93 err/s: 0.00 reconn/s: 0.00
[ 280s ] thds: 64 tps: 471.20 qps: 7541.79 (r/w/o: 6598.80/0.00/943.00) lat (ms,95%): 189.93 err/s: 0.00 reconn/s: 0.00
[ 290s ] thds: 64 tps: 467.80 qps: 7494.69 (r/w/o: 6559.10/0.00/935.60) lat (ms,95%): 196.89 err/s: 0.00 reconn/s: 0.00
[ 300s ] thds: 64 tps: 465.29 qps: 7443.19 (r/w/o: 6512.70/0.00/930.49) lat (ms,95%): 196.89 err/s: 0.00 reconn/s: 0.00
SQL statistics:
    queries performed:
        read:                            1952370
        write:                           0
        other:                           278910
        total:                           2231280
    transactions:                        139455 (464.68 per sec.)
    queries:                             2231280 (7434.89 per sec.)
    ignored errors:                      0      (0.00 per sec.)
    reconnects:                          0      (0.00 per sec.)

General statistics:
    total time:                          300.1067s
    total number of events:              139455

Latency (ms):
         min:                                   56.55
         avg:                                  137.70
         max:                                  287.27
         95th percentile:                      196.89
         sum:                             19202720.10

Threads fairness:
    events (avg/stddev):           2178.9844/5.49
    execution time (avg/stddev):   300.0425/0.03
```